dataGrid to TS
